integration: extract test_unusual_valuelists
Test `test_unusual_valuelists()` is moved from `session/session.rs`
to `types/cql_types.rs` and renamed to
`test_unusual_serializerow_impls()`.
